@@ -1,48 +1,42 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.ComponentModel;
 using System.Diagnostics;
-using System.Drawing;
 using System.Runtime.InteropServices;
 using System.Text;
 using System.Windows;
 using System.Windows.Interop;
 using System.Windows.Media.Imaging;
 
-namespace ShellUtility.Windows.Utility
+namespace ShellUtility.Windows.Utility;
+
+public static class WindowUtility
 {
 
-    public static class WindowUtility
+    #region Pinvoke       
+
+    delegate bool EnumWindowsProc(IntPtr hWnd, IntPtr lParam);
+
+    #region Helpers
+
+    public static IntPtr GetClassLongPtr(IntPtr hWnd, int nIndex) =>
+        IntPtr.Size > 4
+        ? GetClassLongPtr64(hWnd, nIndex)
+        : new IntPtr(GetClassLongPtr32(hWnd, nIndex));
+
+    static bool IsAppWindow(IntPtr hWnd)
     {
 
-        #region Pinvoke       
+        var style = GetWindowLong(hWnd, GWL.GWL_STYLE); // GWL_STYLE
 
-        delegate bool EnumWindowsProc(IntPtr hWnd, IntPtr lParam);
+        // check for WS_VISIBLE and WS_CAPTION flags
+        // (that the window is visible and has a title bar)
+        return (style & 0x10C00000) == 0x10C00000 && (style & 0x10000000) == 0x10000000;
 
-        #region Helpers
+    }
 
-        public static IntPtr GetClassLongPtr(IntPtr hWnd, int nIndex)
-        {
-            if (IntPtr.Size > 4)
-                return GetClassLongPtr64(hWnd, nIndex);
-            else
-                return new IntPtr(GetClassLongPtr32(hWnd, nIndex));
-        }
-
-        static bool IsAppWindow(IntPtr hWnd)
-        {
-
-            var style = GetWindowLong(hWnd, GWL.GWL_STYLE); // GWL_STYLE
-
-            // check for WS_VISIBLE and WS_CAPTION flags
-            // (that the window is visible and has a title bar)
-            return (style & 0x10C00000) == 0x10C00000 && (style & 0x10000000) == 0x10000000;
-
-        }
-
-        const int MaxLastActivePopupIterations = 50;
-        static readonly string[] WindowsClassNamesToSkip =
-        {
+    const int MaxLastActivePopupIterations = 50;
+    static readonly string[] WindowsClassNamesToSkip =
+    {
             "Shell_TrayWnd",
             "DV2ControlHost",
             "MsgrIMEWindowClass",
@@ -50,729 +44,721 @@ namespace ShellUtility.Windows.Utility
             "Button"
         };
 
-        static bool EligibleForActivation(IntPtr hWnd, IntPtr lShellWindow)
+    static bool EligibleForActivation(IntPtr hWnd, IntPtr lShellWindow)
+    {
+
+        // http://stackoverflow.com/questions/210504/enumerate-windows-like-alt-tab-does
+
+        if (hWnd == lShellWindow)
+            return false;
+
+        var root = GetAncestor(hWnd, GetAncestorFlags.GetRootOwner);
+
+        if (GetLastVisibleActivePopUpOfWindow(root) != hWnd)
+            return false;
+
+        var classNameStringBuilder = new StringBuilder(256);
+        var length = GetClassName(hWnd, classNameStringBuilder, classNameStringBuilder.Capacity);
+        if (length == 0)
+            return false;
+
+        var className = classNameStringBuilder.ToString();
+
+        if (Array.IndexOf(WindowsClassNamesToSkip, className) > -1)
+            return false;
+
+        if (className.StartsWith("WMP9MediaBarFlyout")) //WMP's "now playing" taskbar-toolbar
+            return false;
+
+        return true;
+
+    }
+
+    static IntPtr GetLastVisibleActivePopUpOfWindow(IntPtr window)
+    {
+
+        var level = MaxLastActivePopupIterations;
+        var currentWindow = window;
+
+        while (level-- > 0)
         {
+            var lastPopUp = GetLastActivePopup(currentWindow);
 
-            // http://stackoverflow.com/questions/210504/enumerate-windows-like-alt-tab-does
+            if (IsWindowVisible(lastPopUp))
+                return lastPopUp;
 
-            if (hWnd == lShellWindow)
-                return false;
-
-            var root = GetAncestor(hWnd, GetAncestorFlags.GetRootOwner);
-
-            if (GetLastVisibleActivePopUpOfWindow(root) != hWnd)
-                return false;
-
-            var classNameStringBuilder = new StringBuilder(256);
-            var length = GetClassName(hWnd, classNameStringBuilder, classNameStringBuilder.Capacity);
-            if (length == 0)
-                return false;
-
-            var className = classNameStringBuilder.ToString();
-
-            if (Array.IndexOf(WindowsClassNamesToSkip, className) > -1)
-                return false;
-
-            if (className.StartsWith("WMP9MediaBarFlyout")) //WMP's "now playing" taskbar-toolbar
-                return false;
-
-            return true;
-
-        }
-
-        static IntPtr GetLastVisibleActivePopUpOfWindow(IntPtr window)
-        {
-
-            var level = MaxLastActivePopupIterations;
-            var currentWindow = window;
-
-            while (level-- > 0)
-            {
-                var lastPopUp = GetLastActivePopup(currentWindow);
-
-                if (IsWindowVisible(lastPopUp))
-                    return lastPopUp;
-
-                if (lastPopUp == currentWindow)
-                    return IntPtr.Zero;
-
-                currentWindow = lastPopUp;
-            }
-
-            return IntPtr.Zero;
-
-        }
-
-        #endregion
-        #region Methods
-
-        [DllImport("user32.dll", ExactSpelling = true, CharSet = CharSet.Auto)]
-        static extern IntPtr GetParent(IntPtr hWnd);
-
-        [DllImport("user32.dll", SetLastError = false)]
-        static extern IntPtr GetDesktopWindow();
-
-        [DllImport("kernel32.dll", SetLastError = true, CharSet = CharSet.Unicode)]
-        static extern bool QueryFullProcessImageName([In] IntPtr hProcess, [In] int dwFlags, [Out] StringBuilder lpExeName, ref int lpdwSize);
-
-        [DllImport("kernel32.dll", SetLastError = true)]
-        static extern IntPtr OpenProcess(ProcessAccess processAccess, bool bInheritHandle, int processId);
-
-        [DllImport("user32.dll", SetLastError = true)]
-        [return: MarshalAs(UnmanagedType.Bool)]
-        static extern bool GetWindowPlacement(IntPtr hWnd, ref WINDOWPLACEMENT lpwndpl);
-
-        [DllImport("user32.dll")]
-        static extern IntPtr GetWindowThreadProcessId(IntPtr hWnd, out IntPtr ProcessId);
-
-        [DllImport("user32.dll")]
-        static extern IntPtr GetForegroundWindow();
-
-        [DllImport("user32.dll")]
-        [return: MarshalAs(UnmanagedType.Bool)]
-        static extern bool SetForegroundWindow(IntPtr hWnd);
-
-        [DllImport("user32.dll")]
-        [return: MarshalAs(UnmanagedType.Bool)]
-        static extern bool IsWindow(IntPtr hWnd);
-
-        [DllImport("user32.dll", CharSet = CharSet.Unicode, SetLastError = true)]
-        static extern int GetWindowText(IntPtr hWnd, StringBuilder lpString, int nMaxCount);
-
-        [DllImport("user32.dll", SetLastError = true, CharSet = CharSet.Auto)]
-        static extern int GetWindowTextLength(IntPtr hWnd);
-
-        [DllImport("user32.dll", EntryPoint = "GetClassLong")]
-        static extern uint GetClassLongPtr32(IntPtr hWnd, int nIndex);
-
-        [DllImport("user32.dll", EntryPoint = "GetClassLongPtr")]
-        static extern IntPtr GetClassLongPtr64(IntPtr hWnd, int nIndex);
-
-        [DllImport("user32.dll", CharSet = CharSet.Auto, SetLastError = false)]
-        static extern IntPtr SendMessage(IntPtr hWnd, int Msg, int wParam, int lParam);
-        [DllImport("user32.dll", SetLastError = true)]
-        static extern int GetWindowLong(IntPtr hWnd, GWL nIndex);
-
-        [DllImport("user32.dll", ExactSpelling = true)]
-        static extern IntPtr GetAncestor(IntPtr hwnd, GetAncestorFlags flags);
-
-        [DllImport("user32.dll")]
-        static extern IntPtr GetShellWindow();
-
-        [DllImport("user32.dll", SetLastError = true, CharSet = CharSet.Unicode)]
-        static extern int GetClassName(IntPtr hWnd, StringBuilder lpClassName, int nMaxCount);
-
-        [DllImport("user32.dll")]
-        static extern IntPtr GetLastActivePopup(IntPtr hWnd);
-
-        [DllImport("user32.dll")]
-        [return: MarshalAs(UnmanagedType.Bool)]
-        static extern bool IsWindowVisible(IntPtr hWnd);
-
-        [DllImport("user32.dll")]
-        static extern bool EnumDesktopWindows(IntPtr hDesktop, EnumDesktopWindowsDelegate lpfn, IntPtr lParam);
-        delegate bool EnumDesktopWindowsDelegate(IntPtr hWnd, int lParam);
-
-        [DllImport("user32.dll")]
-        [return: MarshalAs(UnmanagedType.Bool)]
-        static extern bool EnumWindows(EnumWindowsProc lpEnumFunc, IntPtr lParam);
-
-        [DllImport("user32.dll")]
-        static extern bool ShowWindowAsync(IntPtr hWnd, SW nCmdShow);
-
-        [DllImport("user32.dll", SetLastError = true)]
-        static extern IntPtr GetWindow(IntPtr hWnd, GetWindowType uCmd);
-
-        [DllImport("user32.dll")]
-        static extern IntPtr LoadIcon(IntPtr hInstance, IntPtr lpIconName);
-
-        #endregion
-        #region Enums and consts
-
-        const int GCL_HICONSM = -34;
-        const int GCL_HICON = -14;
-
-        const int ICON_SMALL = 0;
-        const int ICON_BIG = 1;
-        const int ICON_SMALL2 = 2;
-
-        const int WM_GETICON = 0x7F;
-        const int WM_CLOSE = 0x0010;
-
-        [Flags]
-        public enum ProcessAccess : uint
-        {
-            All = 0x001F0FFF,
-            Terminate = 0x00000001,
-            CreateThread = 0x00000002,
-            VirtualMemoryOperation = 0x00000008,
-            VirtualMemoryRead = 0x00000010,
-            VirtualMemoryWrite = 0x00000020,
-            DuplicateHandle = 0x00000040,
-            CreateProcess = 0x000000080,
-            SetQuota = 0x00000100,
-            SetInformation = 0x00000200,
-            QueryInformation = 0x00000400,
-            QueryLimitedInformation = 0x00001000,
-            Synchronize = 0x00100000
-        }
-
-        enum ShowWindowCommands : int
-        {
-            Hide = 0,
-            Normal = 1,
-            Minimized = 2,
-            Maximized = 3,
-        }
-
-        enum GetWindowType : uint
-        {
-            /// <summary>
-            /// The retrieved handle identifies the window of the same type that is highest in the Z order.
-            /// <para/>
-            /// If the specified window is a topmost window, the handle identifies a topmost window.
-            /// If the specified window is a top-level window, the handle identifies a top-level window.
-            /// If the specified window is a child window, the handle identifies a sibling window.
-            /// </summary>
-            GW_HWNDFIRST = 0,
-            /// <summary>
-            /// The retrieved handle identifies the window of the same type that is lowest in the Z order.
-            /// <para />
-            /// If the specified window is a topmost window, the handle identifies a topmost window.
-            /// If the specified window is a top-level window, the handle identifies a top-level window.
-            /// If the specified window is a child window, the handle identifies a sibling window.
-            /// </summary>
-            GW_HWNDLAST = 1,
-            /// <summary>
-            /// The retrieved handle identifies the window below the specified window in the Z order.
-            /// <para />
-            /// If the specified window is a topmost window, the handle identifies a topmost window.
-            /// If the specified window is a top-level window, the handle identifies a top-level window.
-            /// If the specified window is a child window, the handle identifies a sibling window.
-            /// </summary>
-            GW_HWNDNEXT = 2,
-            /// <summary>
-            /// The retrieved handle identifies the window above the specified window in the Z order.
-            /// <para />
-            /// If the specified window is a topmost window, the handle identifies a topmost window.
-            /// If the specified window is a top-level window, the handle identifies a top-level window.
-            /// If the specified window is a child window, the handle identifies a sibling window.
-            /// </summary>
-            GW_HWNDPREV = 3,
-            /// <summary>
-            /// The retrieved handle identifies the specified window's owner window, if any.
-            /// </summary>
-            GW_OWNER = 4,
-            /// <summary>
-            /// The retrieved handle identifies the child window at the top of the Z order,
-            /// if the specified window is a parent window; otherwise, the retrieved handle is NULL.
-            /// The function examines only child windows of the specified window. It does not examine descendant windows.
-            /// </summary>
-            GW_CHILD = 5,
-            /// <summary>
-            /// The retrieved handle identifies the enabled popup window owned by the specified window (the
-            /// search uses the first such window found using GW_HWNDNEXT); otherwise, if there are no enabled
-            /// popup windows, the retrieved handle is that of the specified window.
-            /// </summary>
-            GW_ENABLEDPOPUP = 6
-        }
-
-        enum SW : int
-        {
-            HIDE = 0,
-            SHOWNORMAL = 1,
-            SHOWMINIMIZED = 2,
-            SHOWMAXIMIZED = 3,
-            SHOWNOACTIVATE = 4,
-            SHOW = 5,
-            MINIMIZE = 6,
-            SHOWMINNOACTIVE = 7,
-            SHOWNA = 8,
-            RESTORE = 9,
-            SHOWDEFAULT = 10
-        }
-
-        enum GWL
-        {
-            GWL_WNDPROC = (-4),
-            GWL_HINSTANCE = (-6),
-            GWL_HWNDPARENT = (-8),
-            GWL_STYLE = (-16),
-            GWL_EXSTYLE = (-20),
-            GWL_USERDATA = (-21),
-            GWL_ID = (-12)
-        }
-
-        enum WindowLongFlags : int
-        {
-            GWL_EXSTYLE = -20,
-            GWLP_HINSTANCE = -6,
-            GWLP_HWNDPARENT = -8,
-            GWL_ID = -12,
-            GWL_STYLE = -16,
-            GWL_USERDATA = -21,
-            GWL_WNDPROC = -4,
-            DWLP_USER = 0x8,
-            DWLP_MSGRESULT = 0x0,
-            DWLP_DLGPROC = 0x4
-        }
-
-        [Flags]
-        public enum WindowStylesEx : uint
-        {
-            /// <summary>Specifies a window that accepts drag-drop files.</summary>
-            WS_EX_ACCEPTFILES = 0x00000010,
-
-            /// <summary>Forces a top-level window onto the taskbar when the window is visible.</summary>
-            WS_EX_APPWINDOW = 0x00040000,
-
-            /// <summary>Specifies a window that has a border with a sunken edge.</summary>
-            WS_EX_CLIENTEDGE = 0x00000200,
-
-            /// <summary>
-            /// Specifies a window that paints all descendants in bottom-to-top painting order using double-buffering.
-            /// This cannot be used if the window has a class style of either CS_OWNDC or CS_CLASSDC. This style is not supported in Windows 2000.
-            /// </summary>
-            /// <remarks>
-            /// With WS_EX_COMPOSITED set, all descendants of a window get bottom-to-top painting order using double-buffering.
-            /// Bottom-to-top painting order allows a descendent window to have translucency (alpha) and transparency (color-key) effects,
-            /// but only if the descendent window also has the WS_EX_TRANSPARENT bit set.
-            /// Double-buffering allows the window and its descendents to be painted without flicker.
-            /// </remarks>
-            WS_EX_COMPOSITED = 0x02000000,
-
-            /// <summary>
-            /// Specifies a window that includes a question mark in the title bar. When the user clicks the question mark,
-            /// the cursor changes to a question mark with a pointer. If the user then clicks a child window, the child receives a WM_HELP message.
-            /// The child window should pass the message to the parent window procedure, which should call the WinHelp function using the HELP_WM_HELP command.
-            /// The Help application displays a pop-up window that typically contains help for the child window.
-            /// WS_EX_CONTEXTHELP cannot be used with the WS_MAXIMIZEBOX or WS_MINIMIZEBOX styles.
-            /// </summary>
-            WS_EX_CONTEXTHELP = 0x00000400,
-
-            /// <summary>
-            /// Specifies a window which contains child windows that should take part in dialog box navigation.
-            /// If this style is specified, the dialog manager recurses into children of this window when performing navigation operations
-            /// such as handling the TAB key, an arrow key, or a keyboard mnemonic.
-            /// </summary>
-            WS_EX_CONTROLPARENT = 0x00010000,
-
-            /// <summary>Specifies a window that has a double border.</summary>
-            WS_EX_DLGMODALFRAME = 0x00000001,
-
-            /// <summary>
-            /// Specifies a window that is a layered window.
-            /// This cannot be used for child windows or if the window has a class style of either CS_OWNDC or CS_CLASSDC.
-            /// </summary>
-            WS_EX_LAYERED = 0x00080000,
-
-            /// <summary>
-            /// Specifies a window with the horizontal origin on the right edge. Increasing horizontal values advance to the left.
-            /// The shell language must support reading-order alignment for this to take effect.
-            /// </summary>
-            WS_EX_LAYOUTRTL = 0x00400000,
-
-            /// <summary>Specifies a window that has generic left-aligned properties. This is the default.</summary>
-            WS_EX_LEFT = 0x00000000,
-
-            /// <summary>
-            /// Specifies a window with the vertical scroll bar (if present) to the left of the client area.
-            /// The shell language must support reading-order alignment for this to take effect.
-            /// </summary>
-            WS_EX_LEFTSCROLLBAR = 0x00004000,
-
-            /// <summary>
-            /// Specifies a window that displays text using left-to-right reading-order properties. This is the default.
-            /// </summary>
-            WS_EX_LTRREADING = 0x00000000,
-
-            /// <summary>
-            /// Specifies a multiple-document interface (MDI) child window.
-            /// </summary>
-            WS_EX_MDICHILD = 0x00000040,
-
-            /// <summary>
-            /// Specifies a top-level window created with this style does not become the foreground window when the user clicks it.
-            /// The system does not bring this window to the foreground when the user minimizes or closes the foreground window.
-            /// The window does not appear on the taskbar by default. To force the window to appear on the taskbar, use the WS_EX_APPWINDOW style.
-            /// To activate the window, use the SetActiveWindow or SetForegroundWindow function.
-            /// </summary>
-            WS_EX_NOACTIVATE = 0x08000000,
-
-            /// <summary>
-            /// Specifies a window which does not pass its window layout to its child windows.
-            /// </summary>
-            WS_EX_NOINHERITLAYOUT = 0x00100000,
-
-            /// <summary>
-            /// Specifies that a child window created with this style does not send the WM_PARENTNOTIFY message to its parent window when it is created or destroyed.
-            /// </summary>
-            WS_EX_NOPARENTNOTIFY = 0x00000004,
-
-            /// <summary>
-            /// The window does not render to a redirection surface.
-            /// This is for windows that do not have visible content or that use mechanisms other than surfaces to provide their visual.
-            /// </summary>
-            WS_EX_NOREDIRECTIONBITMAP = 0x00200000,
-
-            /// <summary>Specifies an overlapped window.</summary>
-            WS_EX_OVERLAPPEDWINDOW = WS_EX_WINDOWEDGE | WS_EX_CLIENTEDGE,
-
-            /// <summary>Specifies a palette window, which is a modeless dialog box that presents an array of commands.</summary>
-            WS_EX_PALETTEWINDOW = WS_EX_WINDOWEDGE | WS_EX_TOOLWINDOW | WS_EX_TOPMOST,
-
-            /// <summary>
-            /// Specifies a window that has generic "right-aligned" properties. This depends on the window class.
-            /// The shell language must support reading-order alignment for this to take effect.
-            /// Using the WS_EX_RIGHT style has the same effect as using the SS_RIGHT (static), ES_RIGHT (edit), and BS_RIGHT/BS_RIGHTBUTTON (button) control styles.
-            /// </summary>
-            WS_EX_RIGHT = 0x00001000,
-
-            /// <summary>Specifies a window with the vertical scroll bar (if present) to the right of the client area. This is the default.</summary>
-            WS_EX_RIGHTSCROLLBAR = 0x00000000,
-
-            /// <summary>
-            /// Specifies a window that displays text using right-to-left reading-order properties.
-            /// The shell language must support reading-order alignment for this to take effect.
-            /// </summary>
-            WS_EX_RTLREADING = 0x00002000,
-
-            /// <summary>Specifies a window with a three-dimensional border style intended to be used for items that do not accept user input.</summary>
-            WS_EX_STATICEDGE = 0x00020000,
-
-            /// <summary>
-            /// Specifies a window that is intended to be used as a floating toolbar.
-            /// A tool window has a title bar that is shorter than a normal title bar, and the window title is drawn using a smaller font.
-            /// A tool window does not appear in the taskbar or in the dialog that appears when the user presses ALT+TAB.
-            /// If a tool window has a system menu, its icon is not displayed on the title bar.
-            /// However, you can display the system menu by right-clicking or by typing ALT+SPACE.
-            /// </summary>
-            WS_EX_TOOLWINDOW = 0x00000080,
-
-            /// <summary>
-            /// Specifies a window that should be placed above all non-topmost windows and should stay above them, even when the window is deactivated.
-            /// To add or remove this style, use the SetWindowPos function.
-            /// </summary>
-            WS_EX_TOPMOST = 0x00000008,
-
-            /// <summary>
-            /// Specifies a window that should not be painted until siblings beneath the window (that were created by the same thread) have been painted.
-            /// The window appears transparent because the bits of underlying sibling windows have already been painted.
-            /// To achieve transparency without these restrictions, use the SetWindowRgn function.
-            /// </summary>
-            WS_EX_TRANSPARENT = 0x00000020,
-
-            /// <summary>Specifies a window that has a border with a raised edge.</summary>
-            WS_EX_WINDOWEDGE = 0x00000100
-        }
-
-        [Flags]
-        public enum WindowStyles : uint
-        {
-            /// <summary>The window has a thin-line border.</summary>
-            WS_BORDER = 0x800000,
-
-            /// <summary>The window has a title bar (includes the WS_BORDER style).</summary>
-            WS_CAPTION = 0xc00000,
-
-            /// <summary>The window is a child window. A window with this style cannot have a menu bar. This style cannot be used with the WS_POPUP style.</summary>
-            WS_CHILD = 0x40000000,
-
-            /// <summary>Excludes the area occupied by child windows when drawing occurs within the parent window. This style is used when creating the parent window.</summary>
-            WS_CLIPCHILDREN = 0x2000000,
-
-            /// <summary>
-            /// Clips child windows relative to each other; that is, when a particular child window receives a WM_PAINT message, the WS_CLIPSIBLINGS style clips all other overlapping child windows out of the region of the child window to be updated.
-            /// If WS_CLIPSIBLINGS is not specified and child windows overlap, it is possible, when drawing within the client area of a child window, to draw within the client area of a neighboring child window.
-            /// </summary>
-            WS_CLIPSIBLINGS = 0x4000000,
-
-            /// <summary>The window is initially disabled. A disabled window cannot receive input from the user. To change this after a window has been created, use the EnableWindow function.</summary>
-            WS_DISABLED = 0x8000000,
-
-            /// <summary>The window has a border of a style typically used with dialog boxes. A window with this style cannot have a title bar.</summary>
-            WS_DLGFRAME = 0x400000,
-
-            /// <summary>
-            /// The window is the first control of a group of controls. The group consists of this first control and all controls defined after it, up to the next control with the WS_GROUP style.
-            /// The first control in each group usually has the WS_TABSTOP style so that the user can move from group to group. The user can subsequently change the keyboard focus from one control in the group to the next control in the group by using the direction keys.
-            /// You can turn this style on and off to change dialog box navigation. To change this style after a window has been created, use the SetWindowLong function.
-            /// </summary>
-            WS_GROUP = 0x20000,
-
-            /// <summary>The window has a horizontal scroll bar.</summary>
-            WS_HSCROLL = 0x100000,
-
-            /// <summary>The window is initially maximized.</summary>
-            WS_MAXIMIZE = 0x1000000,
-
-            /// <summary>The window has a maximize button. Cannot be combined with the WS_EX_CONTEXTHELP style. The WS_SYSMENU style must also be specified.</summary>
-            WS_MAXIMIZEBOX = 0x10000,
-
-            /// <summary>The window is initially minimized.</summary>
-            WS_MINIMIZE = 0x20000000,
-
-            /// <summary>The window has a minimize button. Cannot be combined with the WS_EX_CONTEXTHELP style. The WS_SYSMENU style must also be specified.</summary>
-            WS_MINIMIZEBOX = 0x20000,
-
-            /// <summary>The window is an overlapped window. An overlapped window has a title bar and a border.</summary>
-            WS_OVERLAPPED = 0x0,
-
-            /// <summary>The window is an overlapped window.</summary>
-            WS_OVERLAPPEDWINDOW = WS_OVERLAPPED | WS_CAPTION | WS_SYSMENU | WS_SIZEFRAME | WS_MINIMIZEBOX | WS_MAXIMIZEBOX,
-
-            /// <summary>The window is a pop-up window. This style cannot be used with the WS_CHILD style.</summary>
-            WS_POPUP = 0x80000000u,
-
-            /// <summary>The window is a pop-up window. The WS_CAPTION and WS_POPUPWINDOW styles must be combined to make the window menu visible.</summary>
-            WS_POPUPWINDOW = WS_POPUP | WS_BORDER | WS_SYSMENU,
-
-            /// <summary>The window has a sizing border.</summary>
-            WS_SIZEFRAME = 0x40000,
-
-            /// <summary>The window has a window menu on its title bar. The WS_CAPTION style must also be specified.</summary>
-            WS_SYSMENU = 0x80000,
-
-            /// <summary>
-            /// The window is a control that can receive the keyboard focus when the user presses the TAB key.
-            /// Pressing the TAB key changes the keyboard focus to the next control with the WS_TABSTOP style.  
-            /// You can turn this style on and off to change dialog box navigation. To change this style after a window has been created, use the SetWindowLong function.
-            /// For user-created windows and modeless dialogs to work with tab stops, alter the message loop to call the IsDialogMessage function.
-            /// </summary>
-            WS_TABSTOP = 0x10000,
-
-            /// <summary>The window is initially visible. This style can be turned on and off by using the ShowWindow or SetWindowPos function.</summary>
-            WS_VISIBLE = 0x10000000,
-
-            /// <summary>The window has a vertical scroll bar.</summary>
-            WS_VSCROLL = 0x200000
-        }
-
-        enum GetAncestorFlags
-        {
-            /// <summary>Retrieves the parent window. This does not include the owner, as it does with the GetParent function.</summary>
-            GetParent = 1,
-            /// <summary>Retrieves the root window by walking the chain of parent windows.</summary>
-            GetRoot = 2,
-            /// <summary>Retrieves the owned root window by walking the chain of parent and owner windows returned by GetParent.</summary>
-            GetRootOwner = 3
-        }
-
-        #endregion
-        #region Classes and structs
-
-        [Serializable]
-        [StructLayout(LayoutKind.Sequential)]
-        struct WINDOWPLACEMENT
-        {
-            public int length;
-            public int flags;
-            public ShowWindowCommands showCmd;
-            public System.Drawing.Point ptMinPosition;
-            public System.Drawing.Point ptMaxPosition;
-            public System.Drawing.Rectangle rcNormalPosition;
-        }
-
-        class RECT
-        {
-            public int Left { get; set; }
-            public int Top { get; set; }
-            public int Right { get; set; }
-            public int Bottom { get; set; }
-        }
-
-        #endregion
-
-        #endregion
-
-        public static string GetClassname(IntPtr handle)
-        {
-            var sb = new StringBuilder(256);
-            _ = GetClassName(handle, sb, 256);
-            return sb.ToString();
-        }
-
-        public static bool IsKnownWindowToFilterOut(IntPtr handle) =>
-            GetClassname(handle) == "Windows.UI.Core.CoreWindow" ||
-            (GetProcessAndPath(handle).path.EndsWith("explorer.exe") && string.IsNullOrEmpty(GetTitle(handle)));
-
-        public static bool IsUWPWindow(IntPtr handle)
-        {
-            var classname = GetClassname(handle);
-            return classname == "Windows.UI.Core.CoreWindow" ||
-                   classname == "ApplicationFrameWindow";
-        }
-
-        public static bool IsVisibleInTaskbar(IntPtr handle)
-        {
-
-            if (IsOwned(handle))
-                return false;
-
-            var shellWindow = GetShellWindow();
-
-            var exstyle = (WindowStylesEx)GetWindowLong(handle, GWL.GWL_EXSTYLE);
-
-            var isAppWindow = exstyle.HasFlag(WindowStylesEx.WS_EX_APPWINDOW);
-            var isEligbleForActivation = EligibleForActivation(handle, shellWindow);
-            var isToolWindow = exstyle.HasFlag(WindowStylesEx.WS_EX_TOOLWINDOW);
-
-            if (!isAppWindow && 
-                !(isEligbleForActivation && !isToolWindow))
-                    return false;
-
-            if (IsUWPWindow(handle) && !UWPWindowUtility.IsOpen(handle))
-                return false;
-
-            return true;
-
-        }
-
-        public static (WindowStyles style, WindowStylesEx exStyle) GetWindowStyle(IntPtr handle)
-        {
-
-            var style = (WindowStyles)GetWindowLong(handle, GWL.GWL_STYLE);
-            var exStyle = (WindowStylesEx)GetWindowLong(handle, GWL.GWL_EXSTYLE);
-
-            return (style, exStyle);
-
-        }
-
-        public static bool IsDesktopWindow(IntPtr handle)
-        {
-            
-            if (!IsWindowVisible(handle))
-                return false;
-
-            if (IsKnownWindowToFilterOut(handle))
-                return false;
-
-            return true;
-
-        }
-
-        public static IEnumerable<DesktopWindow> Enumerate()
-        {
-
-            var list = new List<DesktopWindow>();
-            EnumDesktopWindows(IntPtr.Zero, (handle, lParam) =>
-            {
-
-                if (IsDesktopWindow(handle))
-                    list.Add(DesktopWindow.FromHandle(handle));
-                return true;
-
-            }, IntPtr.Zero);
-
-            return list;
-
-        }
-
-        public static bool IsOwned(IntPtr handle) =>
-            GetWindow(handle, GetWindowType.GW_OWNER) != IntPtr.Zero;
-
-        public static IntPtr GetActiveWindow() =>
-            GetForegroundWindow();
-
-        public static bool IsOpen(IntPtr handle) =>
-            IsWindow(handle);
-
-        public static (bool isVisible, Rect? rect) GetIsVisibleAndRect(IntPtr handle)
-        {
-              
-            var placement = new WINDOWPLACEMENT();
-            GetWindowPlacement(handle, ref placement);
-            var isVisible = placement.showCmd != ShowWindowCommands.Minimized;
-
-            if (placement.rcNormalPosition.Width < 0 || placement.rcNormalPosition.Height < 0)
-                return (isVisible, null);
-            else
-                return (isVisible, ToRect(placement.rcNormalPosition));
-
-        }
-
-        static Rect ToRect(Rectangle r) =>
-            new Rect(r.X, r.Y, r.Width, r.Height);
-
-        public static bool IsActive(IntPtr handle) =>
-            GetForegroundWindow() == handle;
-
-        public static void Activate(IntPtr handle) =>
-            SetForegroundWindow(handle);
-
-        public static void Deactivate(IntPtr handle) =>
-            ShowWindowAsync(handle, SW.MINIMIZE);
-
-        public static void SetVisible(IntPtr handle, bool visible)
-        {
-            SetForegroundWindow(handle);
-            ShowWindowAsync(handle, visible ? SW.RESTORE : SW.MINIMIZE);
-        }
-
-        public static void Close(IntPtr handle) =>
-            SendMessage(handle, WM_CLOSE, 0, 0);
-
-        public static string GetTitle(IntPtr handle)
-        {
-
-            var length = GetWindowTextLength(handle);
-            var sb = new StringBuilder(length + 1);
-            if (GetWindowText(handle, sb, 256) != 0)
-                return sb.ToString();
-            else
-                return null;
-        }
-
-        public static IntPtr GetIconHandle(IntPtr handle)
-        {
-
-            if (!IsWindow(handle) || IsUWPWindow(handle))
+            if (lastPopUp == currentWindow)
                 return IntPtr.Zero;
 
-            Func<IntPtr>[] qualities = 
-            { 
+            currentWindow = lastPopUp;
+        }
+
+        return IntPtr.Zero;
+
+    }
+
+    #endregion
+    #region Methods
+
+    [DllImport("user32.dll", ExactSpelling = true, CharSet = CharSet.Auto)]
+    static extern IntPtr GetParent(IntPtr hWnd);
+
+    [DllImport("user32.dll", SetLastError = false)]
+    static extern IntPtr GetDesktopWindow();
+
+    [DllImport("kernel32.dll", SetLastError = true, CharSet = CharSet.Unicode)]
+    static extern bool QueryFullProcessImageName([In] IntPtr hProcess, [In] int dwFlags, [Out] StringBuilder lpExeName, ref int lpdwSize);
+
+    [DllImport("kernel32.dll", SetLastError = true)]
+    static extern IntPtr OpenProcess(ProcessAccess processAccess, bool bInheritHandle, int processId);
+
+    [DllImport("user32.dll", SetLastError = true)]
+    [return: MarshalAs(UnmanagedType.Bool)]
+    static extern bool GetWindowPlacement(IntPtr hWnd, ref WINDOWPLACEMENT lpwndpl);
+
+    [DllImport("user32.dll")]
+    static extern IntPtr GetWindowThreadProcessId(IntPtr hWnd, out IntPtr ProcessId);
+
+    [DllImport("user32.dll")]
+    static extern IntPtr GetForegroundWindow();
+
+    [DllImport("user32.dll")]
+    [return: MarshalAs(UnmanagedType.Bool)]
+    static extern bool SetForegroundWindow(IntPtr hWnd);
+
+    [DllImport("user32.dll")]
+    [return: MarshalAs(UnmanagedType.Bool)]
+    static extern bool IsWindow(IntPtr hWnd);
+
+    [DllImport("user32.dll", CharSet = CharSet.Unicode, SetLastError = true)]
+    static extern int GetWindowText(IntPtr hWnd, StringBuilder lpString, int nMaxCount);
+
+    [DllImport("user32.dll", SetLastError = true, CharSet = CharSet.Auto)]
+    static extern int GetWindowTextLength(IntPtr hWnd);
+
+    [DllImport("user32.dll", EntryPoint = "GetClassLong")]
+    static extern uint GetClassLongPtr32(IntPtr hWnd, int nIndex);
+
+    [DllImport("user32.dll", EntryPoint = "GetClassLongPtr")]
+    static extern IntPtr GetClassLongPtr64(IntPtr hWnd, int nIndex);
+
+    [DllImport("user32.dll", CharSet = CharSet.Auto, SetLastError = false)]
+    static extern IntPtr SendMessage(IntPtr hWnd, int Msg, int wParam, int lParam);
+    [DllImport("user32.dll", SetLastError = true)]
+    static extern int GetWindowLong(IntPtr hWnd, GWL nIndex);
+
+    [DllImport("user32.dll", ExactSpelling = true)]
+    static extern IntPtr GetAncestor(IntPtr hwnd, GetAncestorFlags flags);
+
+    [DllImport("user32.dll")]
+    static extern IntPtr GetShellWindow();
+
+    [DllImport("user32.dll", SetLastError = true, CharSet = CharSet.Unicode)]
+    static extern int GetClassName(IntPtr hWnd, StringBuilder lpClassName, int nMaxCount);
+
+    [DllImport("user32.dll")]
+    static extern IntPtr GetLastActivePopup(IntPtr hWnd);
+
+    [DllImport("user32.dll")]
+    [return: MarshalAs(UnmanagedType.Bool)]
+    static extern bool IsWindowVisible(IntPtr hWnd);
+
+    [DllImport("user32.dll")]
+    static extern bool EnumDesktopWindows(IntPtr hDesktop, EnumDesktopWindowsDelegate lpfn, IntPtr lParam);
+    delegate bool EnumDesktopWindowsDelegate(IntPtr hWnd, int lParam);
+
+    [DllImport("user32.dll")]
+    [return: MarshalAs(UnmanagedType.Bool)]
+    static extern bool EnumWindows(EnumWindowsProc lpEnumFunc, IntPtr lParam);
+
+    [DllImport("user32.dll")]
+    static extern bool ShowWindowAsync(IntPtr hWnd, SW nCmdShow);
+
+    [DllImport("user32.dll", SetLastError = true)]
+    static extern IntPtr GetWindow(IntPtr hWnd, GetWindowType uCmd);
+
+    [DllImport("user32.dll")]
+    static extern IntPtr LoadIcon(IntPtr hInstance, IntPtr lpIconName);
+
+    #endregion
+    #region Enums and consts
+
+    const int GCL_HICONSM = -34;
+    const int GCL_HICON = -14;
+
+    const int ICON_SMALL = 0;
+    const int ICON_BIG = 1;
+    const int ICON_SMALL2 = 2;
+
+    const int WM_GETICON = 0x7F;
+    const int WM_CLOSE = 0x0010;
+
+    [Flags]
+    public enum ProcessAccess : uint
+    {
+        All = 0x001F0FFF,
+        Terminate = 0x00000001,
+        CreateThread = 0x00000002,
+        VirtualMemoryOperation = 0x00000008,
+        VirtualMemoryRead = 0x00000010,
+        VirtualMemoryWrite = 0x00000020,
+        DuplicateHandle = 0x00000040,
+        CreateProcess = 0x000000080,
+        SetQuota = 0x00000100,
+        SetInformation = 0x00000200,
+        QueryInformation = 0x00000400,
+        QueryLimitedInformation = 0x00001000,
+        Synchronize = 0x00100000
+    }
+
+    enum ShowWindowCommands : int
+    {
+        Hide = 0,
+        Normal = 1,
+        Minimized = 2,
+        Maximized = 3,
+    }
+
+    enum GetWindowType : uint
+    {
+        /// <summary>
+        /// The retrieved handle identifies the window of the same type that is highest in the Z order.
+        /// <para/>
+        /// If the specified window is a topmost window, the handle identifies a topmost window.
+        /// If the specified window is a top-level window, the handle identifies a top-level window.
+        /// If the specified window is a child window, the handle identifies a sibling window.
+        /// </summary>
+        GW_HWNDFIRST = 0,
+        /// <summary>
+        /// The retrieved handle identifies the window of the same type that is lowest in the Z order.
+        /// <para />
+        /// If the specified window is a topmost window, the handle identifies a topmost window.
+        /// If the specified window is a top-level window, the handle identifies a top-level window.
+        /// If the specified window is a child window, the handle identifies a sibling window.
+        /// </summary>
+        GW_HWNDLAST = 1,
+        /// <summary>
+        /// The retrieved handle identifies the window below the specified window in the Z order.
+        /// <para />
+        /// If the specified window is a topmost window, the handle identifies a topmost window.
+        /// If the specified window is a top-level window, the handle identifies a top-level window.
+        /// If the specified window is a child window, the handle identifies a sibling window.
+        /// </summary>
+        GW_HWNDNEXT = 2,
+        /// <summary>
+        /// The retrieved handle identifies the window above the specified window in the Z order.
+        /// <para />
+        /// If the specified window is a topmost window, the handle identifies a topmost window.
+        /// If the specified window is a top-level window, the handle identifies a top-level window.
+        /// If the specified window is a child window, the handle identifies a sibling window.
+        /// </summary>
+        GW_HWNDPREV = 3,
+        /// <summary>
+        /// The retrieved handle identifies the specified window's owner window, if any.
+        /// </summary>
+        GW_OWNER = 4,
+        /// <summary>
+        /// The retrieved handle identifies the child window at the top of the Z order,
+        /// if the specified window is a parent window; otherwise, the retrieved handle is NULL.
+        /// The function examines only child windows of the specified window. It does not examine descendant windows.
+        /// </summary>
+        GW_CHILD = 5,
+        /// <summary>
+        /// The retrieved handle identifies the enabled popup window owned by the specified window (the
+        /// search uses the first such window found using GW_HWNDNEXT); otherwise, if there are no enabled
+        /// popup windows, the retrieved handle is that of the specified window.
+        /// </summary>
+        GW_ENABLEDPOPUP = 6
+    }
+
+    enum SW : int
+    {
+        HIDE = 0,
+        SHOWNORMAL = 1,
+        SHOWMINIMIZED = 2,
+        SHOWMAXIMIZED = 3,
+        SHOWNOACTIVATE = 4,
+        SHOW = 5,
+        MINIMIZE = 6,
+        SHOWMINNOACTIVE = 7,
+        SHOWNA = 8,
+        RESTORE = 9,
+        SHOWDEFAULT = 10
+    }
+
+    enum GWL
+    {
+        GWL_WNDPROC = (-4),
+        GWL_HINSTANCE = (-6),
+        GWL_HWNDPARENT = (-8),
+        GWL_STYLE = (-16),
+        GWL_EXSTYLE = (-20),
+        GWL_USERDATA = (-21),
+        GWL_ID = (-12)
+    }
+
+    enum WindowLongFlags : int
+    {
+        GWL_EXSTYLE = -20,
+        GWLP_HINSTANCE = -6,
+        GWLP_HWNDPARENT = -8,
+        GWL_ID = -12,
+        GWL_STYLE = -16,
+        GWL_USERDATA = -21,
+        GWL_WNDPROC = -4,
+        DWLP_USER = 0x8,
+        DWLP_MSGRESULT = 0x0,
+        DWLP_DLGPROC = 0x4
+    }
+
+    [Flags]
+    public enum WindowStylesEx : uint
+    {
+        /// <summary>Specifies a window that accepts drag-drop files.</summary>
+        WS_EX_ACCEPTFILES = 0x00000010,
+
+        /// <summary>Forces a top-level window onto the taskbar when the window is visible.</summary>
+        WS_EX_APPWINDOW = 0x00040000,
+
+        /// <summary>Specifies a window that has a border with a sunken edge.</summary>
+        WS_EX_CLIENTEDGE = 0x00000200,
+
+        /// <summary>
+        /// Specifies a window that paints all descendants in bottom-to-top painting order using double-buffering.
+        /// This cannot be used if the window has a class style of either CS_OWNDC or CS_CLASSDC. This style is not supported in Windows 2000.
+        /// </summary>
+        /// <remarks>
+        /// With WS_EX_COMPOSITED set, all descendants of a window get bottom-to-top painting order using double-buffering.
+        /// Bottom-to-top painting order allows a descendent window to have translucency (alpha) and transparency (color-key) effects,
+        /// but only if the descendent window also has the WS_EX_TRANSPARENT bit set.
+        /// Double-buffering allows the window and its descendents to be painted without flicker.
+        /// </remarks>
+        WS_EX_COMPOSITED = 0x02000000,
+
+        /// <summary>
+        /// Specifies a window that includes a question mark in the title bar. When the user clicks the question mark,
+        /// the cursor changes to a question mark with a pointer. If the user then clicks a child window, the child receives a WM_HELP message.
+        /// The child window should pass the message to the parent window procedure, which should call the WinHelp function using the HELP_WM_HELP command.
+        /// The Help application displays a pop-up window that typically contains help for the child window.
+        /// WS_EX_CONTEXTHELP cannot be used with the WS_MAXIMIZEBOX or WS_MINIMIZEBOX styles.
+        /// </summary>
+        WS_EX_CONTEXTHELP = 0x00000400,
+
+        /// <summary>
+        /// Specifies a window which contains child windows that should take part in dialog box navigation.
+        /// If this style is specified, the dialog manager recurses into children of this window when performing navigation operations
+        /// such as handling the TAB key, an arrow key, or a keyboard mnemonic.
+        /// </summary>
+        WS_EX_CONTROLPARENT = 0x00010000,
+
+        /// <summary>Specifies a window that has a double border.</summary>
+        WS_EX_DLGMODALFRAME = 0x00000001,
+
+        /// <summary>
+        /// Specifies a window that is a layered window.
+        /// This cannot be used for child windows or if the window has a class style of either CS_OWNDC or CS_CLASSDC.
+        /// </summary>
+        WS_EX_LAYERED = 0x00080000,
+
+        /// <summary>
+        /// Specifies a window with the horizontal origin on the right edge. Increasing horizontal values advance to the left.
+        /// The shell language must support reading-order alignment for this to take effect.
+        /// </summary>
+        WS_EX_LAYOUTRTL = 0x00400000,
+
+        /// <summary>Specifies a window that has generic left-aligned properties. This is the default.</summary>
+        WS_EX_LEFT = 0x00000000,
+
+        /// <summary>
+        /// Specifies a window with the vertical scroll bar (if present) to the left of the client area.
+        /// The shell language must support reading-order alignment for this to take effect.
+        /// </summary>
+        WS_EX_LEFTSCROLLBAR = 0x00004000,
+
+        /// <summary>
+        /// Specifies a window that displays text using left-to-right reading-order properties. This is the default.
+        /// </summary>
+        WS_EX_LTRREADING = 0x00000000,
+
+        /// <summary>
+        /// Specifies a multiple-document interface (MDI) child window.
+        /// </summary>
+        WS_EX_MDICHILD = 0x00000040,
+
+        /// <summary>
+        /// Specifies a top-level window created with this style does not become the foreground window when the user clicks it.
+        /// The system does not bring this window to the foreground when the user minimizes or closes the foreground window.
+        /// The window does not appear on the taskbar by default. To force the window to appear on the taskbar, use the WS_EX_APPWINDOW style.
+        /// To activate the window, use the SetActiveWindow or SetForegroundWindow function.
+        /// </summary>
+        WS_EX_NOACTIVATE = 0x08000000,
+
+        /// <summary>
+        /// Specifies a window which does not pass its window layout to its child windows.
+        /// </summary>
+        WS_EX_NOINHERITLAYOUT = 0x00100000,
+
+        /// <summary>
+        /// Specifies that a child window created with this style does not send the WM_PARENTNOTIFY message to its parent window when it is created or destroyed.
+        /// </summary>
+        WS_EX_NOPARENTNOTIFY = 0x00000004,
+
+        /// <summary>
+        /// The window does not render to a redirection surface.
+        /// This is for windows that do not have visible content or that use mechanisms other than surfaces to provide their visual.
+        /// </summary>
+        WS_EX_NOREDIRECTIONBITMAP = 0x00200000,
+
+        /// <summary>Specifies an overlapped window.</summary>
+        WS_EX_OVERLAPPEDWINDOW = WS_EX_WINDOWEDGE | WS_EX_CLIENTEDGE,
+
+        /// <summary>Specifies a palette window, which is a modeless dialog box that presents an array of commands.</summary>
+        WS_EX_PALETTEWINDOW = WS_EX_WINDOWEDGE | WS_EX_TOOLWINDOW | WS_EX_TOPMOST,
+
+        /// <summary>
+        /// Specifies a window that has generic "right-aligned" properties. This depends on the window class.
+        /// The shell language must support reading-order alignment for this to take effect.
+        /// Using the WS_EX_RIGHT style has the same effect as using the SS_RIGHT (static), ES_RIGHT (edit), and BS_RIGHT/BS_RIGHTBUTTON (button) control styles.
+        /// </summary>
+        WS_EX_RIGHT = 0x00001000,
+
+        /// <summary>Specifies a window with the vertical scroll bar (if present) to the right of the client area. This is the default.</summary>
+        WS_EX_RIGHTSCROLLBAR = 0x00000000,
+
+        /// <summary>
+        /// Specifies a window that displays text using right-to-left reading-order properties.
+        /// The shell language must support reading-order alignment for this to take effect.
+        /// </summary>
+        WS_EX_RTLREADING = 0x00002000,
+
+        /// <summary>Specifies a window with a three-dimensional border style intended to be used for items that do not accept user input.</summary>
+        WS_EX_STATICEDGE = 0x00020000,
+
+        /// <summary>
+        /// Specifies a window that is intended to be used as a floating toolbar.
+        /// A tool window has a title bar that is shorter than a normal title bar, and the window title is drawn using a smaller font.
+        /// A tool window does not appear in the taskbar or in the dialog that appears when the user presses ALT+TAB.
+        /// If a tool window has a system menu, its icon is not displayed on the title bar.
+        /// However, you can display the system menu by right-clicking or by typing ALT+SPACE.
+        /// </summary>
+        WS_EX_TOOLWINDOW = 0x00000080,
+
+        /// <summary>
+        /// Specifies a window that should be placed above all non-topmost windows and should stay above them, even when the window is deactivated.
+        /// To add or remove this style, use the SetWindowPos function.
+        /// </summary>
+        WS_EX_TOPMOST = 0x00000008,
+
+        /// <summary>
+        /// Specifies a window that should not be painted until siblings beneath the window (that were created by the same thread) have been painted.
+        /// The window appears transparent because the bits of underlying sibling windows have already been painted.
+        /// To achieve transparency without these restrictions, use the SetWindowRgn function.
+        /// </summary>
+        WS_EX_TRANSPARENT = 0x00000020,
+
+        /// <summary>Specifies a window that has a border with a raised edge.</summary>
+        WS_EX_WINDOWEDGE = 0x00000100
+    }
+
+    [Flags]
+    public enum WindowStyles : uint
+    {
+        /// <summary>The window has a thin-line border.</summary>
+        WS_BORDER = 0x800000,
+
+        /// <summary>The window has a title bar (includes the WS_BORDER style).</summary>
+        WS_CAPTION = 0xc00000,
+
+        /// <summary>The window is a child window. A window with this style cannot have a menu bar. This style cannot be used with the WS_POPUP style.</summary>
+        WS_CHILD = 0x40000000,
+
+        /// <summary>Excludes the area occupied by child windows when drawing occurs within the parent window. This style is used when creating the parent window.</summary>
+        WS_CLIPCHILDREN = 0x2000000,
+
+        /// <summary>
+        /// Clips child windows relative to each other; that is, when a particular child window receives a WM_PAINT message, the WS_CLIPSIBLINGS style clips all other overlapping child windows out of the region of the child window to be updated.
+        /// If WS_CLIPSIBLINGS is not specified and child windows overlap, it is possible, when drawing within the client area of a child window, to draw within the client area of a neighboring child window.
+        /// </summary>
+        WS_CLIPSIBLINGS = 0x4000000,
+
+        /// <summary>The window is initially disabled. A disabled window cannot receive input from the user. To change this after a window has been created, use the EnableWindow function.</summary>
+        WS_DISABLED = 0x8000000,
+
+        /// <summary>The window has a border of a style typically used with dialog boxes. A window with this style cannot have a title bar.</summary>
+        WS_DLGFRAME = 0x400000,
+
+        /// <summary>
+        /// The window is the first control of a group of controls. The group consists of this first control and all controls defined after it, up to the next control with the WS_GROUP style.
+        /// The first control in each group usually has the WS_TABSTOP style so that the user can move from group to group. The user can subsequently change the keyboard focus from one control in the group to the next control in the group by using the direction keys.
+        /// You can turn this style on and off to change dialog box navigation. To change this style after a window has been created, use the SetWindowLong function.
+        /// </summary>
+        WS_GROUP = 0x20000,
+
+        /// <summary>The window has a horizontal scroll bar.</summary>
+        WS_HSCROLL = 0x100000,
+
+        /// <summary>The window is initially maximized.</summary>
+        WS_MAXIMIZE = 0x1000000,
+
+        /// <summary>The window has a maximize button. Cannot be combined with the WS_EX_CONTEXTHELP style. The WS_SYSMENU style must also be specified.</summary>
+        WS_MAXIMIZEBOX = 0x10000,
+
+        /// <summary>The window is initially minimized.</summary>
+        WS_MINIMIZE = 0x20000000,
+
+        /// <summary>The window has a minimize button. Cannot be combined with the WS_EX_CONTEXTHELP style. The WS_SYSMENU style must also be specified.</summary>
+        WS_MINIMIZEBOX = 0x20000,
+
+        /// <summary>The window is an overlapped window. An overlapped window has a title bar and a border.</summary>
+        WS_OVERLAPPED = 0x0,
+
+        /// <summary>The window is an overlapped window.</summary>
+        WS_OVERLAPPEDWINDOW = WS_OVERLAPPED | WS_CAPTION | WS_SYSMENU | WS_SIZEFRAME | WS_MINIMIZEBOX | WS_MAXIMIZEBOX,
+
+        /// <summary>The window is a pop-up window. This style cannot be used with the WS_CHILD style.</summary>
+        WS_POPUP = 0x80000000u,
+
+        /// <summary>The window is a pop-up window. The WS_CAPTION and WS_POPUPWINDOW styles must be combined to make the window menu visible.</summary>
+        WS_POPUPWINDOW = WS_POPUP | WS_BORDER | WS_SYSMENU,
+
+        /// <summary>The window has a sizing border.</summary>
+        WS_SIZEFRAME = 0x40000,
+
+        /// <summary>The window has a window menu on its title bar. The WS_CAPTION style must also be specified.</summary>
+        WS_SYSMENU = 0x80000,
+
+        /// <summary>
+        /// The window is a control that can receive the keyboard focus when the user presses the TAB key.
+        /// Pressing the TAB key changes the keyboard focus to the next control with the WS_TABSTOP style.  
+        /// You can turn this style on and off to change dialog box navigation. To change this style after a window has been created, use the SetWindowLong function.
+        /// For user-created windows and modeless dialogs to work with tab stops, alter the message loop to call the IsDialogMessage function.
+        /// </summary>
+        WS_TABSTOP = 0x10000,
+
+        /// <summary>The window is initially visible. This style can be turned on and off by using the ShowWindow or SetWindowPos function.</summary>
+        WS_VISIBLE = 0x10000000,
+
+        /// <summary>The window has a vertical scroll bar.</summary>
+        WS_VSCROLL = 0x200000
+    }
+
+    enum GetAncestorFlags
+    {
+        /// <summary>Retrieves the parent window. This does not include the owner, as it does with the GetParent function.</summary>
+        GetParent = 1,
+        /// <summary>Retrieves the root window by walking the chain of parent windows.</summary>
+        GetRoot = 2,
+        /// <summary>Retrieves the owned root window by walking the chain of parent and owner windows returned by GetParent.</summary>
+        GetRootOwner = 3
+    }
+
+    #endregion
+    #region Classes and structs
+
+    [Serializable]
+    [StructLayout(LayoutKind.Sequential)]
+    struct WINDOWPLACEMENT
+    {
+        public int length;
+        public int flags;
+        public ShowWindowCommands showCmd;
+        public POINT ptMinPosition;
+        public POINT ptMaxPosition;
+        public RECT rcNormalPosition;
+    }
+
+    [StructLayout(LayoutKind.Sequential)]
+    class RECT
+    {
+        public int Left { get; set; }
+        public int Top { get; set; }
+        public int Right { get; set; }
+        public int Bottom { get; set; }
+    }
+
+    [StructLayout(LayoutKind.Sequential)]
+    public struct POINT
+    {
+        public int X { get; set; }
+        public int Y { get; set; }
+    }
+
+    #endregion
+
+    #endregion
+
+    public static string GetClassname(IntPtr handle)
+    {
+        var sb = new StringBuilder(256);
+        _ = GetClassName(handle, sb, 256);
+        return sb.ToString();
+    }
+
+    public static bool IsKnownWindowToFilterOut(IntPtr handle) =>
+        GetClassname(handle) == "Windows.UI.Core.CoreWindow" ||
+        (GetProcessAndPath(handle).path.EndsWith("explorer.exe") && string.IsNullOrEmpty(GetTitle(handle)));
+
+    public static bool IsUWPWindow(IntPtr handle) =>
+        GetClassname(handle) is "Windows.UI.Core.CoreWindow" or "ApplicationFrameWindow";
+
+    public static bool IsVisibleInTaskbar(IntPtr handle)
+    {
+
+        if (IsOwned(handle))
+            return false;
+
+        var shellWindow = GetShellWindow();
+
+        var exstyle = (WindowStylesEx)GetWindowLong(handle, GWL.GWL_EXSTYLE);
+
+        var isAppWindow = exstyle.HasFlag(WindowStylesEx.WS_EX_APPWINDOW);
+        var isEligbleForActivation = EligibleForActivation(handle, shellWindow);
+        var isToolWindow = exstyle.HasFlag(WindowStylesEx.WS_EX_TOOLWINDOW);
+
+        if (!isAppWindow &&
+            !(isEligbleForActivation && !isToolWindow))
+            return false;
+
+        if (IsUWPWindow(handle) && !UWPWindowUtility.IsOpen(handle))
+            return false;
+
+        return true;
+
+    }
+
+    public static (WindowStyles style, WindowStylesEx exStyle) GetWindowStyle(IntPtr handle)
+    {
+
+        var style = (WindowStyles)GetWindowLong(handle, GWL.GWL_STYLE);
+        var exStyle = (WindowStylesEx)GetWindowLong(handle, GWL.GWL_EXSTYLE);
+
+        return (style, exStyle);
+
+    }
+
+    public static bool IsDesktopWindow(IntPtr handle) =>
+        IsWindowVisible(handle) && !IsKnownWindowToFilterOut(handle);
+
+    public static IEnumerable<DesktopWindow> Enumerate()
+    {
+
+        var list = new List<DesktopWindow>();
+        _ = EnumDesktopWindows(IntPtr.Zero, (handle, lParam) =>
+          {
+
+              if (IsDesktopWindow(handle))
+                  list.Add(DesktopWindow.FromHandle(handle));
+              return true;
+
+          }, IntPtr.Zero);
+
+        return list;
+
+    }
+
+    public static bool IsOwned(IntPtr handle) =>
+        GetWindow(handle, GetWindowType.GW_OWNER) != IntPtr.Zero;
+
+    public static IntPtr GetActiveWindow() =>
+        GetForegroundWindow();
+
+    public static bool IsOpen(IntPtr handle) =>
+        IsWindow(handle);
+
+    public static (bool isVisible, Rect? rect) GetIsVisibleAndRect(IntPtr handle)
+    {
+
+        var placement = new WINDOWPLACEMENT();
+        _ = GetWindowPlacement(handle, ref placement);
+        var isVisible =
+            placement.showCmd is ShowWindowCommands.Normal or ShowWindowCommands.Maximized &&
+            GetWindowStyle(handle).style.HasFlag(WindowStyles.WS_VISIBLE);
+
+        return (isVisible, ToRect(placement.rcNormalPosition));
+
+    }
+
+    static Rect ToRect(RECT r) =>
+        new(r.Left, r.Top, r.Right - r.Left, r.Bottom - r.Top);
+
+    public static bool IsActive(IntPtr handle) =>
+        GetForegroundWindow() == handle;
+
+    public static void Activate(IntPtr handle) =>
+        SetForegroundWindow(handle);
+
+    public static void Deactivate(IntPtr handle) =>
+        ShowWindowAsync(handle, SW.MINIMIZE);
+
+    public static void SetVisible(IntPtr handle, bool visible)
+    {
+        _ = SetForegroundWindow(handle);
+        _ = ShowWindowAsync(handle, visible ? SW.RESTORE : SW.MINIMIZE);
+    }
+
+    public static void Close(IntPtr handle) =>
+        SendMessage(handle, WM_CLOSE, 0, 0);
+
+    public static string GetTitle(IntPtr handle)
+    {
+
+        var length = GetWindowTextLength(handle);
+        var sb = new StringBuilder(length + 1);
+        return
+            GetWindowText(handle, sb, 256) != 0
+            ? sb.ToString()
+            : null;
+
+    }
+
+    public static IntPtr GetIconHandle(IntPtr handle)
+    {
+
+        if (!IsWindow(handle) || IsUWPWindow(handle))
+            return IntPtr.Zero;
+
+        Func<IntPtr>[] qualities =
+        {
                 () => GetClassLongPtr(handle, GCL_HICON),
                 () => SendMessage(handle, WM_GETICON, ICON_BIG, 0),
                 () => SendMessage(handle, WM_GETICON, ICON_SMALL2, 0),
             };
 
-            foreach (var method in qualities)
-                if (method?.Invoke() is IntPtr hIcon && hIcon != IntPtr.Zero)
-                    return hIcon;
+        foreach (var method in qualities)
+            if (method?.Invoke() is IntPtr hIcon && hIcon != IntPtr.Zero)
+                return hIcon;
 
-            return LoadIcon(IntPtr.Zero, (IntPtr)0x7F00/*IDI_APPLICATION*/);
+        return LoadIcon(IntPtr.Zero, (IntPtr)0x7F00/*IDI_APPLICATION*/);
 
-        }
+    }
 
-        public static BitmapSource GetIcon(IntPtr iconHandle)
-        {
+    public static BitmapSource GetIcon(IntPtr iconHandle)
+    {
 
-            if (iconHandle == IntPtr.Zero)
-                return null;
+        if (iconHandle == IntPtr.Zero)
+            return null;
 
-            var image = Imaging.CreateBitmapSourceFromHIcon(iconHandle, Int32Rect.Empty, BitmapSizeOptions.FromWidthAndHeight(26, 26));
-            image.Freeze();
-            return image;
+        var image = Imaging.CreateBitmapSourceFromHIcon(iconHandle, Int32Rect.Empty, BitmapSizeOptions.FromWidthAndHeight(26, 26));
+        image.Freeze();
+        return image;
 
-        }
+    }
 
-        public static (Process process, string path) GetProcessAndPath(IntPtr handle)
-        {
+    public static (Process process, string path) GetProcessAndPath(IntPtr handle)
+    {
 
-            if (handle == IntPtr.Zero)
-                return default;
+        if (handle == IntPtr.Zero)
+            return default;
 
-            GetWindowThreadProcessId(handle, out var pid);
-            var process = Process.GetProcessById((int)pid);
+        _ = GetWindowThreadProcessId(handle, out var pid);
+        var process = Process.GetProcessById((int)pid);
 
-            var processHandle = OpenProcess(ProcessAccess.QueryLimitedInformation, false, process.Id);
-            var capacity = 1024;
-            var path = new StringBuilder(capacity);
-            QueryFullProcessImageName(processHandle, 0, path, ref capacity);
+        var processHandle = OpenProcess(ProcessAccess.QueryLimitedInformation, false, process.Id);
+        var capacity = 1024;
+        var path = new StringBuilder(capacity);
+        _ = QueryFullProcessImageName(processHandle, 0, path, ref capacity);
 
-            return (process, path.ToString());
-
-        }
+        return (process, path.ToString());
 
     }
 
